requirements.txt
Flask==2.3.2
PyJWT==2.6.0
bcrypt==4.0.1
cryptography==38.0.4
psycopg2-binary==2.9.10
sqlalchemy
pydantic==2.7.1
gunicorn==20.1.0